feat: extract DetailOrder component from Checkout
